first draft for cleaning up state file
